// server/backend/src/types/address.ts
type Address = {
        first_name?: string
        address1?: string
        phone?: string
        city?: string
        zip?: string
        province?: string
        country?: string
        last_name?: string
        address2?: string
        company?: string
        latitude?: string
        longitude?: string
        name?: string
        country_code?: string
        province_code?: string
}

export default Address